docs: Add guidance on installing the programming env
This commit updates the README file to include information on installing
the necessary programming packages for users with and without Anaconda.
I also included necessary commands for upgrading Anaconda, creating the
environment from the .yaml file, activating the environment, and opening
the jupyter notebook.  I also included links to download Anaconda and
links to the python programming packages I use during the project.