feat: Change "hypothesis" to "expectation" in Intro
I changed the word hypothesis to expectation in the Introduction section
of the main jupyter notebook file (Test a Perceptual Phenomenon.ipynb).
I did this so that there wouldn't be confusion between the word hypothesis
in the next section (Definition of Independent and Dependent Variables and
Hypotheses) and in the Introduction.