feat: Add descriptive statistics of sample data
This commit adds descriptive statistics to the jupyter notebook for the
sample data to determine which condition has higher mean and median
response times only for this sample data.  A statistical test still
needs to be performed to say something about the entire human
population.  Also added is a measure of variability (the standard
deviation) to get an idea of the spread of the data and references for
each line of code  so that contributers can see where I determined
how to perform each calculation.