feat: Add visualization of distributions
This commit adds visualizations of the distributions of times to
correctly state the color of ink under each condition for the
sample data.  I did this to visually compare the data under each
condition.  I wanted to visually show that in general, the times
for the incongruent condition are greater than that of the
congruent condition.  I also added references which explained
how to adjust each portion of the histogram.