docs: Edit README to reflect updated testing type
Again, I realized that the type of t-test I was originally doing
was wrong.  I was assuming the data was independent when in fact
the data is dependent since the same subjects are giving a time
for both the congruent and incongruent conditions.  Because of
this, I am having to go back and edit each section to reflect
this change and realization.  I will now be using a paired
two-sample t-test.  This commit edits the Intro in the README.