feat: Edit visualizations to match updated testing
I realized that the type of t-test I was originally doing was wrong.
I was assuming the data was independent when in fact the data is
dependent since the same subjects are giving a time for both the
congruent and incongruent conditions.  Because of this, I am having
to go back and edit each section to reflect this change and
realization.  I will now be using a paired two-sample t-test. This
commit edits the section on visualization of the distribution of
differences such that now I show the distribution of difference in
time to state the correct color between each condition.  I also
updated the references and the Resources section to reflect these
changes.