feat: Define mu correctly in section on hypotheses
I realized I defined mu wrong in the section on hyptheses. I said mu
was the mean across all subjects when in fact it is the mean across
the entire population (however that's defined).  The population
could be the entire human population or a subset but something much
larger than the sample of 24 subjects we have data for.