feat: Add performance of hypothesis test & implications
This commit adds the hypothesis test for the two-sample, paired, right-
tailed t-test.  I also explained what the t-test was and how the
p-value and t-statistic helps us determine whether to reject the null
hypothesis.  I added what rejecting the null hypothesis meant for
the difference in times between the incongruent and congruent
conditions on average.  Finally, I added any relevant references
to the resources section.